refactor: planet service 통합
