+++ src/main/java/com/planetrush/planetrush/planet/service/PlanetService.java
@@ -2,13 +2,23 @@ package com.planetrush.planetrush.planet.service;
 
 import java.util.List;
 
+import com.planetrush.planetrush.planet.service.dto.GetDefaultPlanetImgDto;
 import com.planetrush.planetrush.planet.service.dto.GetMainPlanetListDto;
 import com.planetrush.planetrush.planet.service.dto.GetMyPlanetListDto;
 import com.planetrush.planetrush.planet.service.dto.OngoingPlanetDto;
 import com.planetrush.planetrush.planet.service.dto.PlanetDetailDto;
+import com.planetrush.planetrush.planet.service.dto.PlanetSubscriptionDto;
+import com.planetrush.planetrush.planet.service.dto.RegisterPlanetDto;
 import com.planetrush.planetrush.planet.service.dto.SearchCond;
 
-public interface GetPlanetService {
+public interface PlanetService {
+
+	/**
+	 * 모든 기본 행성 이미지 URL을 조회합니다.
+	 * @return 모든 기본 행성 이미지의 URL 목록을 가져와 DefaultPlanetImgDto로 반환합니다.
+	 * @see GetDefaultPlanetImgDto
+	 */
+	List<GetDefaultPlanetImgDto> getAllImgUrls();
 
 	/**
 	 * 챌린지가 시작되지 않은 행성을 검색합니다.
@@ -50,4 +60,21 @@ public interface GetPlanetService {
 	 */
 	List<GetMainPlanetListDto> getMainPlanetList(Long memberId);
 
+	/**
+	 * 주어진 dto를 사용하여 입주 기록을 생성합니다.
+	 * @param dto 사용자의 id, 행성의 id 등을 포함합니다.
+	 */
+	void registerResident(PlanetSubscriptionDto dto);
+
+	/**
+	 * 주어진 dto를 사용하여 입주 기록을 삭제합니다.
+	 * @param dto 사용자의 id, 행성의 id 등을 포함합니다.
+	 */
+	void deleteResident(PlanetSubscriptionDto dto);
+
+	/**
+	 * 행성을 등록합니다.
+	 * @param dto 행성 정보
+	 */
+	void registerPlanet(RegisterPlanetDto dto);
 }
